Made various fixes based on feedback from pull request
