--- src/duckling/controls/layer-dialog.component.scss
@@ -1,3 +1,0 @@
-.show-hide-layer-button {
-
-}
