+++ src/duckling/entitysystem/entity-system.module.ts
@@ -11,6 +11,7 @@ import {
     EntitySystemService,
     EntityLayerService
 } from './index';
+import {LayerDialogComponent} from './services/layer-dialog.component';
 
 @NgModule({
     providers: [
@@ -22,6 +23,18 @@ import {
         EntityLayerService,
         BaseAttributeService,
         EntitySystemService
+    ],
+    imports: [
+        ControlsModule
+    ],
+    declarations: [
+        LayerDialogComponent
+    ],
+    exports: [
+        LayerDialogComponent
+    ],
+    entryComponents: [
+        LayerDialogComponent
     ]
 })
 export class EntitySystemModule {
